fetch multiple contacts from google contacts and use the old receiver choice to select between these
