+++ lib/automate/modules/schedule.py
@@ -6,7 +6,7 @@ import logging
 from lib import Error
 from lib.automate.google import Google
 from lib.automate.modules import Module, NoSenderError
-from lib.utils.contacts import get_emails
+from lib.utils.contacts import *
 from datetime import datetime, timedelta
 
 # Module logger
@@ -62,12 +62,8 @@ class Schedule(Module):
         for (name, candidates) in parsed_attendees["uncertain"]:
             self.uncertain_attendee = (name, candidates)
             self.followup_type = "to_uncertain"
-            followup_str = f"Found multiple contacts with the name {name}\n"
-            for i in range(len(candidates)):
-                c_name, c_email = candidates[i]
-                followup_str += f"[{i+1}] {c_name} - {c_email}\n"
-            followup_str += f"Please choose one (1-{len(candidates)})"
-            return followup_str
+            return prompt_contact_choice(name, candidates)
+
         for name in parsed_attendees["unknown"]:
             self.followup_type = "to_unknown"
             self.unknown_attendee = name
@@ -121,7 +117,9 @@ class Schedule(Module):
             except Exception:
                 return self.prepare_processed(self.to, self.when, self.body, self.sender)
             name, candidates = self.uncertain_attendee
-            if choice >= 0 and choice < len(candidates):
+            if choice < 0:
+                raise NoContactFoundError("No contact with name " + name + " was found")
+            elif choice >= 0 and choice < len(candidates):
                 self.to.remove(name)  # update to so recursive call continues resolving new attendees
                 self.to.append(candidates[choice][1])  # add email of chosen attendee
             return self.prepare_processed(self.to, self.when, self.body, self.sender)
